checkpoint - migrate to using long/Long to represent any integer ID value (rowId, objectId, etc).

+++ elisa/src/org/labkey/elisa/ElisaController.java
@@ -55,17 +55,17 @@ public class ElisaController extends SpringActionController
 
     public static class RunDetailsForm
     {
-        private int _protocolId;
-        private int _runId;
+        private long _protocolId;
+        private long _runId;
         private String _runName;
         private String _schemaName;
 
-        public int getRunId()
+        public long getRunId()
         {
             return _runId;
         }
 
-        public void setRunId(int runId)
+        public void setRunId(long runId)
         {
             _runId = runId;
         }
@@ -90,12 +90,12 @@ public class ElisaController extends SpringActionController
             _schemaName = schemaName;
         }
 
-        public int getProtocolId()
+        public long getProtocolId()
         {
             return _protocolId;
         }
 
-        public void setProtocolId(int protocolId)
+        public void setProtocolId(long protocolId)
         {
             _protocolId = protocolId;
         }
